Refactor pt2
