+++ package.json
@@ -13,6 +13,7 @@
    },
    "dependencies": {
       "@grammyjs/auto-retry": "^1.1.1",
+      "@grammyjs/commands": "^1.2.0",
       "@grammyjs/menu": "^1.2.0",
       "@grammyjs/ratelimiter": "^1.1.6",
       "@grammyjs/runner": "^1.0.4",
