+++ src/features/admin/admin.services.ts
@@ -1,3 +1,4 @@
+import { GroupDTO } from 'features/group/group.dto';
 import { IAdminRepository, IAdminServices } from './admin.interfaces';
 
 import { err, ok } from 'shared/result';
@@ -8,7 +9,7 @@ export default class AdminServices implements IAdminServices {
     
     public async getAdminGroups(userId: string) {		
         
-        const response = await this.adminRepository.getAdminWithGroups(userId);
+        const response = await this.adminRepository.getWithGroups(userId);
 
         if(response.ok === true) {
             const groups = response.value.groups;
@@ -24,10 +25,10 @@ export default class AdminServices implements IAdminServices {
         }
     }
     
-    public async editGroupPermissions(groupId: string, userId: string, permissions: object) {
+    public async editGroupPermissions(groupId: string, userId: string, permissions: Partial<GroupDTO>) {
 
         //1) Fetch the admin with their groups
-        const adminResult = await this.adminRepository.getAdminWithGroups(userId);
+        const adminResult = await this.adminRepository.getWithGroups(userId);
 
         if(adminResult.ok === false) {
             if(adminResult.error === "DB_ERROR") {
@@ -43,7 +44,7 @@ export default class AdminServices implements IAdminServices {
             }
 
             //If found, proceed to edit permissions
-            const groupResult = await this.adminRepository.editGroupPermissions(groupId, userId, permissions);
+            const groupResult = await this.adminRepository.editGroupPermissions(groupId, permissions);
             if(groupResult.ok === false) {
                 if(groupResult.error === "DB_ERROR") {
                     return err("INTERNAL_ERROR");
@@ -53,6 +54,45 @@ export default class AdminServices implements IAdminServices {
                 return ok(null);
             }
         }
+    }
+
+    public async addAdmins(groupId: string, userIds: string[]) {
+        const response = await this.adminRepository.addAdmins(groupId, userIds);
+        if(response.ok === false) {
+            return err("INTERNAL_ERROR");
+        }
+        return ok(null);
+    }
+
+    public async removeAdmins(groupId: string, userIds: string[]) {
+        const response = await this.adminRepository.deleteAdmins(groupId, userIds);
+        if(response.ok === false) {
+            return err("INTERNAL_ERROR");
+        }
+        return ok(null);
+    }
+
+    public async reloadAdmins(groupId: string, userIds: string[]) {
+        const deleteResponse = await this.adminRepository.deleteAllAdmins(groupId);
+        if(deleteResponse.ok === false) {
+            return err("INTERNAL_ERROR");
+        }
+
+        const addResponse = await this.adminRepository.addAdmins(groupId, userIds);
+        if(addResponse.ok === false) {
+            return err("INTERNAL_ERROR");
+        }
+
+        return ok(null);
+    }
+
+    public async deleteAllAdmins(groupId: string) {
+        const response = await this.adminRepository.deleteAllAdmins(groupId);
+        if(response.ok === false) {
+            return err("INTERNAL_ERROR");
+        }
+        return ok(null);
+
     }
 }
